feat(ai): promote rust tagger to production and remove python-related files
