+++ apps/server/src/routes/manager.tsx
@@ -19,16 +19,16 @@ import {
 	deleteIp,
 	updateIp,
 } from "~/infrastructure/api-clients/ips-api";
+import {
+	deleteMedia,
+	findDuplicateMedia,
+} from "~/infrastructure/api-clients/media-api";
 import { orpc } from "~/infrastructure/api-clients/orpc-client";
 import {
 	createProject,
 	deleteProject,
 	updateProject,
 } from "~/infrastructure/api-clients/projects-api";
-import {
-	deleteMedia,
-	findDuplicateMedia,
-} from "~/infrastructure/api-clients/media-api";
 import { allCharactersQueryOptions } from "~/infrastructure/api-clients/queries/characters-query";
 import { allIpsQueryOptions } from "~/infrastructure/api-clients/queries/ips-query";
 import { allProjectsQueryOptions } from "~/infrastructure/api-clients/queries/projects-query";
