+++ apps/tauri/src/components/media/media-sidebar/media-sidebar-content.tsx
@@ -26,7 +26,7 @@ import {
 import { buildMediaContentUrl } from "~/infrastructure/media/thumbnail-runtime";
 import { AiTaggingModal } from "../ai-tagging-modal";
 import { CharacterCropModal } from "../character-crop-modal";
-import { RustExperimentalModal } from "../rust-experimental-modal";
+
 
 type MediaSidebarProps = {
 	media: MediaDetails;
@@ -72,13 +72,7 @@ export function MediaSidebar(props: MediaSidebarProps) {
 			removeCharacterFromMedia={removeCharacterFromMedia}
 			removeIpFromMedia={removeIpFromMedia}
 			removeProjectFromMedia={removeProjectFromMedia}
-			rustExperimentalModal={(modalProps) => (
-				<RustExperimentalModal
-					isOpen={modalProps.isOpen}
-					media={props.media}
-					onClose={modalProps.onClose}
-				/>
-			)}
+
 			characterCropModal={(modalProps) => (
 				<CharacterCropModal
 					isOpen={modalProps.isOpen}
