+++ package.json
@@ -12,7 +12,6 @@
     "dev": "bun --filter @solid-imager/server dev",
     "build": "bun --filter '*' build",
     "start": "bun --filter @solid-imager/server start",
-    "ai:start": "uv run uvicorn src-python.main:app --host 0.0.0.0 --port 8000",
     "lint": "vp run lint",
     "check": "vp run check",
     "format": "vp run format",
