--- pyproject.toml
@@ -1,13 +1,0 @@
-[project]
-name = "solid-imager"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.13"
-dependencies = [
-    "dghs-imgutils>=0.19.0",
-    "fastapi>=0.115.0",
-    "uvicorn>=0.32.0",
-    "python-multipart>=0.0.12",
-    "onnxruntime>=1.23.2",
-]
